refactor: reorganize file analysis agent
