+++ file_analysis_agent/__init__.py
@@ -1,5 +1,5 @@
 """File analysis agent package."""
-from .config import AgentConfig, CONFIG, update_config
+from .agent_tools.config import AgentConfig, CONFIG, update_config
 from .agent_tools import tools
 
 __all__ = ["AgentConfig", "CONFIG", "update_config", "tools"]
