+++ file_analysis_agent/agent_tools/cache.py
