+++ pyproject.toml
@@ -5,6 +5,7 @@ description = "Tools for converting documents to Markdown and inspecting them"
 requires-python = ">=3.12"
 dependencies = [
     "pydantic>=2",
+    "pydantic-ai",
     "tiktoken",
     "diskcache",
 ]
